// src/pages/customer/motorbike-component/model/Motorbike.ts
export interface IMotorbike {
    id: number;
    image: string;
    images: string[];
    name: string;
    brand: string;
    model: string;
    vehicleCapacity: number;
    price: number;
    status: string;
    storeName: string;
    yearRegister: Date;
    postDate: Date;
    motorType: string;
    odo: number;
    description: string;
}